src/phase1_capture/chatbot_web/causal_survey/requirements.txt
openai==1.12.0
httpx==0.27.2
pyro-ppl==1.8.6
networkx==3.2.1
matplotlib==3.8.2
pyvis==0.3.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
torch==2.1.2
numpy==1.26.3
pandas==2.1.4 